Initial version (not complete)
For now, the solver only goes through the board looking for unambiguous answers to cells: it will fill in cells for which only one number is possible. It does a few iterations of this until the board no longer changes anymore.

If, after these iterations, there are still empty cells, this means that these cells have multiple possible numbers that can be filled in. We need to develop the logic to go further.